move generic stack into a common type package
