+++ internal/capability/types.go
@@ -2,7 +2,6 @@ package capability
 
 import (
 	"context"
-	"errors"
 	"time"
 
 	"github.com/opdev/opcap/internal/operator"
@@ -27,37 +26,3 @@ type (
 	auditFn        func(context.Context) error
 	auditCleanupFn func(context.Context) error
 )
-
-type Stack[T any] struct {
-	stack *element[T]
-}
-
-type element[T any] struct {
-	previous *element[T]
-	val      T
-}
-
-var StackEmptyError = errors.New("Stack empty")
-
-func (s *Stack[T]) Push(v T) {
-	e := &element[T]{
-		previous: s.stack,
-		val:      v,
-	}
-	s.stack = e
-}
-
-func (s *Stack[T]) Pop() (T, error) {
-	if s.stack == nil {
-		var r T
-		return r, StackEmptyError
-	}
-	e := *s.stack
-	s.stack = e.previous
-
-	return e.val, nil
-}
-
-func (s *Stack[T]) Empty() bool {
-	return s.stack == nil
-}
